+++ Templates/editorTemplate.js
@@ -64,3 +64,36 @@ export const menuPath = ""
 
 /** @type {EditorExecute} */
 export const execute = () => {}
+
+/** @type {OnTick} */
+export const onTick = () => {}
+
+/** @type {OnTestMap} */
+export const onBeforeTestMap = () => {}
+
+/** @type {OnTestMap} */
+export const onAfterTestMap = () => {}
+
+/** @type {OnExportMap} */
+export const onBeforeExportMap = filePath => {}
+
+/** @type {OnExportMap} */
+export const onAfterExportMap = filePath => {}
+
+/** @type {OnOpenScene} */
+export const onBeforeOpenScene = scenePath => {}
+
+/** @type {OnOpenScene} */
+export const onAfterOpenScene = scenePath => {}
+
+/** @type {OnSaveScene} */
+export const onBeforeSaveScene = scenePath => {}
+
+/** @type {OnSaveScene} */
+export const onAfterSaveScene = scenePath => {}
+
+/** @type {OnImportAsset} */
+export const onBeforeImportAsset = (filePath, assetPath) => {}
+
+/** @type {OnImportAsset} */
+export const onAfterImportAsset = (filePath, assetPath) => {}