+++ _Typings/editorApi/exportFuncs.d.ts
@@ -2,9 +2,13 @@
 // For details: https://github.com/Withered-Flower-0422/BST/blob/main/LICENSE
 
 declare module "editor:type" {
+    import { Path } from "path"
+
     global {
-        /** Executed when button is pressed. */
-        type EditorExecute = () => void
+        type EditorExecute = {
+            /** Executed when button is pressed. */
+            (): void
+        }
 
         /**
          * The name displayed in the menu bar for the registered script. Use `/` to separate levels.
@@ -17,5 +21,55 @@ declare module "editor:type" {
          * then place it in the `Scripts/_Editor/Icons` folder and name it as the `menuPath` string in the script.
          */
         type MenuPath = string
+
+        type OnTick = {
+            /** Called every 1 second. */
+            (): void
+        }
+
+        type OnTestMap = {
+            /** Called when testing the map. */
+            (): void
+        }
+
+        type OnExportMap = {
+            /**
+             * Called when exporting the map.
+             * @param filePath The system path of the exported map file.
+             */
+            (filePath: string): void
+        }
+
+        type OnOpenScene = {
+            /**
+             * Called when opening the scene.
+             * @param scenePath The BME path of the scene.
+             */
+            (scenePath: Path<"Scenes", true, false>): void
+        }
+
+        type OnSaveScene = {
+            /**
+             * Called when saving the scene.
+             * @param scenePath The BME path of the scene.
+             */
+            (scenePath: Path<"Scenes", true, false>): void
+        }
+
+        type OnImportAsset = {
+            /**
+             * Called when importing an asset.
+             * @param filePath The system path of the asset.
+             * @param assetPath The BME path of the asset.
+             */
+            (
+                filePath: string,
+                assetPath: Path<
+                    "Meshes" | "Textures" | "Materials" | "Audios" | "Scripts",
+                    true,
+                    false
+                >,
+            ): void
+        }
     }
 }